# Fix binds logging for "WHERE ... IN ..." statements