--- a/activerecord/lib/active_record/statement_cache.rb
+++ b/activerecord/lib/active_record/statement_cache.rb
@@ -80,8 +80,8 @@ module ActiveRecord
         self
       end
 
-      def add_binds(binds)
-        @binds.concat binds
+      def add_binds(binds, proc_for_binds = nil)
+        @binds.concat proc_for_binds ? binds.map(&proc_for_binds) : binds
         binds.size.times do |i|
           @parts << ", " unless i == 0
           @parts << Substitute.new