--- a/activerecord/lib/arel/collectors/bind.rb
+++ b/activerecord/lib/arel/collectors/bind.rb
@@ -16,8 +16,8 @@ module Arel # :nodoc: all
         self
       end
 
-      def add_binds(binds)
-        @binds.concat binds
+      def add_binds(binds, proc_for_binds = nil)
+        @binds.concat proc_for_binds ? binds.map(&proc_for_binds) : binds
         self
       end