@ricardotk002 2021-01-22
Fair point. I'm also unsure if it's the right solution.

I checked #41136 and it seems like a different approach, although with some corner cases:

```ruby
Developer.where(id: [1, 2, 3, 4, 5], created_at: Time.now).load

# New format
# => SELECT "developers".* FROM "developers" WHERE "developers"."id" IN ($1, $2, $3, $4, $5) AND "developers"."created_at" = $6  [1, 2, 3, 4, 5, ["created_at", "2021-01-22 01:49:07.085656"]]

# Previous format
# => SELECT "developers".* FROM "developers" WHERE "developers"."id" IN ($1, $2, $3, $4, $5) AND "developers"."created_at" = $6  [["id", 1], ["id", 2], ["id", 3], ["id", 4], ["id", 5], ["created_at", "2021-01-22 01:50:47.720448"]]
```

On a side note, I ran my initial benchmarks against the baseline you used (`#where` with sanitized SQL). I tried yours and still the difference seems reasonable. 

Before:

```
Warming up --------------------------------------
      where with ids    10.000  i/100ms
 where with sanitize    17.000  i/100ms
Calculating -------------------------------------
      where with ids    115.849  (± 1.7%) i/s -    580.000  in   5.008217s
 where with sanitize    173.993  (± 3.4%) i/s -    884.000  in   5.088834s

Comparison:
 where with sanitize:      174.0 i/s
      where with ids:      115.8 i/s - 1.50x  (± 0.00) slower
```

After:

```
Warming up --------------------------------------
      where with ids    11.000  i/100ms
 where with sanitize    16.000  i/100ms
Calculating -------------------------------------
      where with ids    116.501  (± 1.7%) i/s -    583.000  in   5.006341s
 where with sanitize    176.277  (± 2.3%) i/s -    896.000  in   5.085439s

Comparison:
 where with sanitize:      176.3 i/s
      where with ids:      116.5 i/s - 1.51x  (± 0.00) slower
```

My guess is that the performance gain comes from not having to visit an `In` node and its left/right leaves. The binds are still visited with `#add_binds`, which works similarly to `#inject_join`.

<details>
  <summary>Benchmark script</summary>

```ruby
require "active_record"
require "minitest/autorun"
require "stackprof"
require "benchmark/ips"

ActiveRecord::Base.establish_connection(adapter: "mysql2", username: "rails", database: "activerecord_unittest")

ActiveRecord::Schema.define do
  create_table :posts, force: true do |t|
  end
end

class Post < ActiveRecord::Base
end

class BugTest < Minitest::Test
  def test_association_stuff
    ids = (1..1000).each.map do |n|
      Post.create!.id
    end

    Benchmark.ips do |x|
      x.report("where with ids") do
        Post.where(id: ids).to_a
      end

      x.report("where with sanitize") do
        Post.where(ActiveRecord::Base.sanitize_sql(["id IN (?)", ids])).to_a
      end

      x.compare!
    end
  end
end
```

</details>

Feel free to close this one if the solution doesn't apply, I understand it's up to you to decide which option is better long term.