@ricardotk002 2021-02-23
@eileencodes @kamipo I tried another approach, which is similar (slightly lighter) to the implementation pre-refactor. Performance seems to stay similar using the benchmarks provided above. I also borrowed @kamipo's tests to verify that everything works fine.

The thing you need to determine is if you still want to keep the old logging format (there are some issues that reference this PR):

```
#  SELECT * FROM events WHERE id IN (?, ?)  [["id", 1], ["id", 2]]
```

Or keep the new concise version https://github.com/rails/rails/pull/41136 (related https://github.com/rails/rails/pull/29944)

```
#  SELECT * FROM events WHERE id IN (?, ?)  [1, 2]
```

Which has some curious cases as I mentioned here https://github.com/rails/rails/pull/41068#issuecomment-765066261.

I'm more than happy to try another approach, if you give me some direction. 

Otherwise, I think you should move ahead with the new concise format, I don't want this PR to keep the issue unsolved.