@kamipo 2021-02-23
I'd prefer to make bind attribute only if the bind collector is used, it doesn't have performance penalty if `prepared_statements: false` which is by default for mysql2 adapter.

https://github.com/rails/rails/compare/fix_binds_logging