@kamipo 2021-03-19
I'm merging this since in general changes for the performance should not break any public behavior.

We haven't tested the format of the binds part, but I think losing the attr name from the log for the performance is over optimizing.

Now, changes in this PR makes bind attribute only if the bind collector is used, that is to say it doesn't have performance penalty if `prepared_statements: false` which is by default for mysql2 adapter.